feat: Support for i<C-BS> added.
`<C-BS>` deletes back a word in most programs, but this was not
supported in nvim, which has the same functionality on `<C-w>` by
default. Now either can be used.

Also fixed a minor keymap bug where formatting in insert mode with
`<C-f>` would move the cursor backwards.

Lastly, added a fancy header to the file.