refactor!: reworked keys module (now keymaps)
Complerely revamped the keys module so now it's much more versatile to
use. It works be defining namespace tables on the module which are
arrays of keymap tables. Then you can call setup with the name of that
namespace and optionally a bufnr, and it will set those keymaps for that
buffer. Can also remove keymaps from a namespace and buffer with the
module.