refactor: yicks color scheme module
Now added a proper yicks color scheme module, which just gets required
in the colorscheme file. Internals are a lot cleaner and are now
data-based rather than procedure-based. This makes it a lot easier to
change a minimal number of values for dramatic changes in the output
scheme. Yicks now supports light mode with `'background'`. Because of
the internal changes, this requires *no* additional configuration, it
just works!