fix: shell and lsp options
So apparently setting 'fileignorecase' to false messes with language
servers on windows for some reason? It varies based on platform, so
basically the lesson is to not touch it.

Shell commands were acting funky trying to remove the builtin powershell
`Tee-Object` to use vim's `tee`. The fix was to not try to
`Remove-Alias` but instead to just call `tee.exe`.