feat: Toc command and gO for markdown files
Added a :Toc command which opens a loclist with the table of contents of
a markdown file, formatted using extmarks. Also added an autocmd to make
gO work in markdown files by opening this buffer.

Also removed the global DirContains function because vim.fs.root() does
what that did but better in nearly every way.