feat: updated lines to match new contour updates
Lines now works with the new contour updates, and everything is setup in
a single lazy `opts` table, as it should be. Yicks added custom
highlight groups for contour's diagnostics component, because that
wasn't being highlighted correctly.

Also changed around the contour arrangement a bit, so now the tabline
shows the cwd and time on the left, buffers open in the current tab in
the center, and tab pages by number on the right. The statusline shows
diagnostics on the left, buffer info in the middle, and position on the
right.