refactor: keymaps module
Keymaps module now works more clearly on the basis of keymap namespaces.
Keymaps are added and removed from namespaces, and namespaces are bound
to buffers. When keymaps are added they are also added to all existing
buffers and when tehy are removed they are removed from those buffers as
well.