refactor: removed mason dependancy
The config no longer requires mason. Mason is essentially a package
manager for dev tools, but those tools are usually already packaged for
the platform anyways, so mason isn't needed. I'll just improve the lsp
code once I get rid of the manager dependency to notify the user when a
server is missing.