feat: changed appearance and defaulted options
This started with wanting to hide the sign column as much as possible,
because it's really wide and usually blank. This meant changing the
diagnostic setup to use virtual text instead of signs, but left the
column filled by line numbers. This was also unnecessarily wide, so I
turned that off too. Several other options which weren't doing anything
significant were also defaulted.