refactor!: fcache is now obejct-oriented
This must be the most breaking change ever. All of the old `fcache` api
is now deleted, and replaced with a lovely reference and obejct based
replacement. The changes to fex are just to make it compatible with the
new changes, and also a few small bug fixes which happened during the
update.