refactor: updated lines module
In theory, it is less likely to break now. I've set it up so that there
are only two functions which generate the status line and tab line
respectively, and are defined globally in lua. This should make
following errors much easier. The new method relies more on the built-in
vim status line features and simplified logic.