feat: surrounding with the `sa` operator
Added an operator which activates on the keymap `sa`, or on the whole
line with `saa`. Once fully invoked (with motion / selection), it will
wait for the user to press a key, which it uses to identify the type of
surround. The `t` and `T` (tag) and `p` and `P` (prompt) keys will ask
the user for input which it will surround with.