fix(pairs): change/delete error and block surround
When changing and deleting there was a bug which caused an error to be
printed if the opening character was not found, but nothing should have
happened. Set a default value which doesn't break anything to avoid
`#nil` breaking things.

Block surrounds were inserting the closing character two characters
after the end of the selection, instead of one character after the end
of the selection. Just adjusted the offsets to fix this.