fix(lsp): completion acts on typed character
The previous completion method would trigger on `InsertCharPre`, which
would mean that the typed character would not actually be involved in
completion. The new method uses `TextChangedI` which is far more
reliable in this aspect. Also updated the module so shortmess will
prevent insert completion from causing flickering in the command line,
and changed completion item kind highlight group names.