fix(lsp): completion trigger is more sane
Apparently cancelling or selecting a completion option triggers
TextChangedI for some godforsaken reason. Why on earth would cancelling
completion and doing nothing trigger an event? Who knows. The fix was
checking if completion should be triggered on InsertCharPre, which is
more reasonable, then actually triggering it on TextChangedI so that the
language server actually accounts for the character which was typed (so
it's no longer "Pre").