fix(fex): circular file movement bug
The major bug fix of this commit is an issue where circularly renaming
files in fex (renaming a to b and b to a) would fail miserably, and not
warn the user. This has been fixed to work correctly.

Also fixed a minor issue where the cursor placement auto-correction
would fail in an empty buffer.