fix(fex): fixed issues with reopneing fex buffers
When reopening fex buffers with either `:e` or bringing them back to a
window there were a multitude of errors, all of which were related to
the concealing of the ids. The changes to fix this were to remove the
syntax file, delete the old extmarks, and change the window settings
updates.