feat(calendir): full year calendar
The `:Calendir` command can now take an integer as a year, and the
plugin will open a buffer which displays every day in that year as
columns under the months. Using the `gf` keymap on any of the dates will
jump to that journal entry. Days with journal entries and the current
date are also given special highlight groups.