refactor(scratch): extracted scratch to a plugin
The scratch functionality was taking up a *lot* of space in `init.lua`,
so I decided to extract it to another buffer. This was also done with
the intent of eventually adding the functionality to run scratch buffers
like scripts, though that might be difficult with some languages like
rust which require a project.