feat(pick): added a versatile picker plugin
Added a picker plugin with extensible infrastructure. Currently, it
supports mostly asynchronous grep functionality using ripgrep. I don't
like the default settings when it's used in 'grepprg', and the semantics
of the quickfix commands make it hard to respect them here. It will just
be easier to hard code ripgrep and implement a quickfic/location list
importer later on.