fix(calendir): fixed command and Dec/Jan calendars
The next and previous subcommds used to read the buffer name and extract
the date from there, which was honestly a little dumb. Now, they don't
do that and the plugin instead stores the date and type of buffer in
buffer variables.

The December and January calendars used to have incorrect loop
conditions so that opening any December would cause an infinite loop,
and January would display as empty.