feat(fex)!: changed keymaps
Changed the `<cr>` keymap which opens the file under the cursor to 'gf',
because that's more in line with what vim does by default. Also added an
override for `cc`, which normally changes an entire line. This behavior
isn't optimal in a fex buffer, so now it just removes the file name and
leaves the invisible metadata as is. Extmarks now also disappear when
their lines are deleted.