feat(pairs)!: dot-repeat works better with surround
Updated when the surround commands ask the user for which surround type
should be executed, which means that surrounds can be dot-repeated
without re-typing the surrounding characters every time. Also simplified
the module so that all of the pairs are in a single table, this
unfortunately removed several pairs, but they might be added back later.