feat(pairs): text object surrounds
The keymaps to change and delete surrounds now use text objects to
identify and change the surroundings, which is much better than the
previous method. It has a few quirks, but those quirks closely align
with the quirks of text objects, so for someone familiar with vim they
aren't a big deal.