doc: updated README and helptags
I forgot to update the helptags, so I did that. I also removed the giant
list at the bottom of README because I nearly everything was checked or
not going to happen. The bottom of the file now contains a list of
external plugins and justification for depending on them.