fix(projects): missing projects.json
There was no proper error handling for loading the projects file when it
is missing, so errors would get displayed on startup. The user should
probably not be aware of the projects file, so notifying them of errors
ever is buggy behavior (which existed because `assert()` is easy). The
fix was to just assume there are no saved projects on load errors.